Add links to Shout Out and Call Out pages
These links will link to files I'm going to go create after this, but are crucial to the communal accountability tracking records / score boards. Not to sound too creepy, like that Black Mirror episode where social scores dictated whether or not a person was able to use a service or obtain a good. Direct punishment for life's choices. 

I think this is a softer way to go about this. At the end of the day, I fully expect some people to call me out. I intend to call others out,, but I rewrote a public message several times before trying to help out someone I wasn't a fan of. Even with a pfp of Charlie Kirk (I was organizing colleagues to peacefully debate Charlie when he came to CSU).